rename geom-utils to geometry
