+++ cocos/core/geometry/curve.ts
