+++ cocos/core/index.ts
@@ -27,7 +27,7 @@
  * @category core
  */
 
-import * as geometry from './geom-utils';
+import * as geometry from './geometry';
 import * as math from './math';
 import * as memop from './memop';
 cc.math = math;
